requirements.txt
# MQTT
paho-mqtt>=2.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
asyncpg>=0.29.0

# Monitoring & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# Data Processing
numpy>=1.24.0
scikit-learn>=1.3.0

# Redis
redis>=5.0.0

